Extract "className" assertion to own file
