+++ lib/referee.js
@@ -5,7 +5,6 @@ var interpolatePosArg = require("./interpolate-pos-arg");
 var interpolateProperties = require("./interpolate-properties");
 var expect = require("./expect");
 var bane = require("bane");
-var includes = require("lodash.includes");
 var Promise = require("es6-promise").Promise;
 
 var slice = Array.prototype.slice;
@@ -205,40 +204,7 @@ require("./assertions/near")(referee);
 require("./assertions/has-prototype")(referee);
 require("./assertions/contains")(referee);
 require("./assertions/tag-name")(referee);
-
-referee.add("className", {
-    assert: function (element, name) {
-        if (typeof element.className === "undefined") {
-            return this.fail("noClassNameMessage");
-        }
-
-        var expected = typeof name === "string" ? name.split(" ") : name;
-        var actual = element.className.split(" ");
-        var i, l;
-        for (i = 0, l = expected.length; i < l; i++) {
-            if (!includes(actual, expected[i])) { return false; }
-        }
-
-        return true;
-    },
-    assertMessage: "${customMessage}Expected object's className to include ${expected} " +
-        "but was ${actual}",
-    refuteMessage: "${customMessage}Expected object's className not to include ${expected}",
-    expectation: "toHaveClassName",
-    values: function (element, className, message) {
-        return {
-            actualElement: element,
-            actual: element.className,
-            expected: className,
-            customMessage: message
-        };
-    }
-});
-
-referee.assert.className.noClassNameMessage = "${customMessage}Expected object to have " +
-    "className property";
-referee.refute.className.noClassNameMessage = "${customMessage}Expected object to have " +
-    "className property";
+require("./assertions/class-name")(referee);
 
 referee.expect = function () {
     expect.init(referee);
